Rewrite algorithm to find start/stop for 1x60 spawnpoints
Previous algorithm used some kind of O(n^2) which became a huge problem
for n=1000+. Now it will loop through all pokemons that were sighted 2
or more times (for 1x60 spawnpoints, we can't determine start/end times
with pokemons that were only scanned once). This way we can for each
found pokemon expand the start/end window and that way find the range
in O(n).

Unit tests were added and as a result the args variable had to be moved
from the root level of imports.